+++ kaye/cli/cli_main.py
@@ -13,7 +13,7 @@ from kaye import PROGRAM_NAME
 
 from kaye.cli.cli_http import register_cli_http_parser
 from kaye.cli.cli_continue import register_cli_continue_parser
-from kaye.cli.cli_skill.cli_skill_main import register_cli_skill_parser
+from kaye.cli.cli_claude.cli_claude_main import register_cli_claude_parser
 
 __all__ = ("cli_parser", "cli_subparser")
 
@@ -31,4 +31,4 @@ cli_subparser = cli_parser.add_subparsers(title="subcommands")
 # register subcommands parsers
 register_cli_continue_parser(cli_subparser)
 register_cli_http_parser(cli_subparser)
-register_cli_skill_parser(cli_subparser)
+register_cli_claude_parser(cli_subparser)